Specify TargetRubyVersion
In RuboCop's default, TargetRubyVersion is 2.0.
But we almost use the higher version.
If TargetRubyVersion is older than used ruby version, RuboCop says
`unexpected token ...` It's a false positive detection. So, I've specify
the newest ruby version.

Problem
-------

In the setting, RuboCop may detects false negative when use older ruby
version than TargetRubyVersion. But I think it is not important.